requirements.txt
rembg==2.0.65
Pillow==10.3.0
onnxruntime==1.21.0 